notion_search_app/requirements.txt
flask
python-dotenv
requests
notion-client
openai
google-search-results
gunicorn
